auth working and notifications working, but need to move to redux to manage state better
